model modif

+++ fire/map/models.py
@@ -66,8 +66,8 @@ class Data(models.Model):
     IdData = models.AutoField(primary_key=True,default=None)
     temperature = models.BigIntegerField(null=True)
     humidity = models.BigIntegerField(null=True)
-    wind = models.BigIntegerField(default= 0,null=True)
-    rain = models.BigIntegerField(default= 0,null=True)
+    wind = models.FloatField(default=0, null=True)
+    rain = models.FloatField(default=0, null=True)
     node = models.ForeignKey(node, on_delete=models.CASCADE, null=True, related_name='datas')
